+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## 6.?.? (2019-11-??)
+## 6.2.0 (2019-11-??)
 
-* Add more details to exception messages when storing or loading an image fails
+* Added support for jpeg images with new `stubbles\img\driver\JpegDriver`
+* Added more details to exception messages when storing or loading an image fails
 
 ## 6.1.0 (2019-11-18)